Rename cli_modules to modules, it's cleaner
