+++ python/fedml/cli/modules/device.py
